n_win_k must be an integer
Otherwhise NumPy complains:

    DeprecationWarning: using a non-integer number instead of an integer
will result in an error in the future
    y[0]=np.mean(y[0:n_win])

This uncovered a small bug in mod_filter_picker.py, where there is an
integer division (1/n_win_mb) which would provide 0 as result.

+++ BackProj.py
@@ -76,7 +76,7 @@ dt=st[0].stats.delta
 fs_data = st[0].stats.sampling_rate
 dT=dt
 npts_d = st[0].stats.npts
-n_win_k = config.decay_const /dt
+n_win_k = int(config.decay_const/dt)
 
 st_CF=st.copy()
